requirements.txt
requests
urllib3
beautifulsoup4
python-dotenv
colorama
cryptography
PgQueuer
sqlalchemy[asyncio]
psycopg2-binary
asyncpg
fake-useragent
numpy
Pillow
opencv-python
fastapi
uvicorn
httpx
python-multipart
python-telegram-bot
curl_cffi
ddddocr
apscheduler
pydantic-settings
playwright
pytest
pytest-asyncio
psutil
